requirements.txt
# MCP Framework
mcp==1.0.0

# HTTP Clients for external APIs
httpx>=0.27
requests==2.31.0

# Google Places API
googlemaps==4.10.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1

# For SQLite (development)
# For PostgreSQL (production) - uncomment if needed
# psycopg2-binary==2.9.9

# Configuration & Environment
pydantic>=2.8.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Caching
diskcache>=5.6.3

# Logging
structlog>=23.2.0

# Data Processing
geopy>=2.4.1

# Date & Time utilities
python-dateutil>=2.8.2

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0